import printErrors as a util method
