+++ lib/index.js
@@ -1,5 +1,6 @@
 const logger = require('./logger')
 const schemaBuilder = require('./schema_builder')
 const utils = require('./utils')
+const printErrors  = require('print_errors')
 
-module.exports = { logger, schemaBuilder, utils }
+module.exports = { logger, schemaBuilder, utils, printErrors }
