+++ src/index.js
@@ -1,6 +1,6 @@
-const { logger, schemaBuilder, utils } = require('../lib')
 const fs = require('fs')
 const Schema = require('validate')
+const { logger, schemaBuilder, utils, printErrors } = require('../lib')
 
 const logLevels = {
   none: 0,
@@ -58,24 +58,5 @@ const validateExtraFields = (targetObj, schemaObj) => {
   _parseTarget(targetObj, schemaObj)
   return extras;
 }
-
-const printErrors = (errors, warnings, logLevel = 3) => {
-  if(errors.length || warnings.length) {
-    if(logLevel > 2) {
-      logger.error('====== Schema Validation Error ======')
-      logger.error(`${errors.length} mismatches and ${warnings.length} warnings found.`)
-    }
-    if(logLevel) {
-      errors.forEach((err, index) => logger.red(`${index + 1}. ${err.message}`))
-    }
-    if(logLevel > 1) {
-      warnings.forEach((warn, index) => logger.yellow(`${index + 1}. ${warn.message}`))
-    }
-  } else if(logLevel > 2) {
-    logger.success('Schema Validated Successfully')
-  }
-  return errors.concat(warnings)
-}
-
 
 module.exports = validateSchema
